+++ db/db_api.go
@@ -11,6 +11,7 @@ import (
 	sparql "github.com/gtfierro/hod/lang/ast"
 	"github.com/gtfierro/hod/turtle"
 	logrus "github.com/sirupsen/logrus"
+	"github.com/syndtr/goleveldb/leveldb"
 
 	"github.com/blevesearch/bleve"
 	"github.com/coocood/freecache"
@@ -513,6 +514,7 @@ func (db *DB) getQueryResults(q *sparql.Query) ([]*ResultRow, queryStats, error)
 
 	runStart := time.Now()
 	ctx, err := db.executeQueryPlan(qp)
+	defer ctx.Close()
 	if err != nil {
 		return nil, stats, err
 	}
@@ -528,7 +530,10 @@ func (db *DB) getQueryResults(q *sparql.Query) ([]*ResultRow, queryStats, error)
 }
 
 func (db *DB) executeQueryPlan(plan *queryPlan) (*queryContext, error) {
-	ctx := newQueryContext(plan, db)
+	ctx, err := newQueryContext(plan, db)
+	if err != nil {
+		return nil, errors.Wrap(err, "Could not get snapshot")
+	}
 
 	for _, op := range ctx.operations {
 		now := time.Now()
@@ -560,4 +565,193 @@ func (db *DB) sortQueryTerms(q *sparql.Query) *dependencyGraph {
 
 	dg.terms = terms
 	return dg
+}
+
+// returns the uint64 hash of the given URI (this is adjusted for uniqueness)
+func (db *DB) GetHash(entity turtle.URI) (Key, error) {
+	var rethash Key
+	if hash, err := db.entityHashCache.Get(entity.Bytes()); err != nil {
+		if err == freecache.ErrNotFound {
+			val, err := db.entityDB.Get(entity.Bytes(), nil)
+			if err != nil {
+				return emptyKey, errors.Wrapf(err, "Could not get Entity for %s", entity)
+			}
+			copy(rethash[:], val)
+			if rethash == emptyKey {
+				return emptyKey, errors.New("Got bad hash")
+			}
+			db.entityHashCache.Set(entity.Bytes(), rethash[:], -1) // no expiry
+			return rethash, nil
+		} else {
+			return emptyKey, errors.Wrapf(err, "Could not get Entity for %s", entity)
+		}
+	} else {
+		copy(rethash[:], hash)
+	}
+	return rethash, nil
+}
+
+func (db *DB) MustGetHash(entity turtle.URI) Key {
+	val, err := db.GetHash(entity)
+	if err != nil {
+		log.Error(errors.Wrapf(err, "Could not get hash for %s", entity))
+		return emptyKey
+	}
+	return val
+}
+
+func (db *DB) GetURI(hash Key) (turtle.URI, error) {
+	db.uriLock.RLock()
+	if uri, found := db.uriCache[hash]; found {
+		db.uriLock.RUnlock()
+		return uri, nil
+	}
+	db.uriLock.RUnlock()
+	db.uriLock.Lock()
+	defer db.uriLock.Unlock()
+	val, err := db.pkDB.Get(hash[:], nil)
+	if err != nil {
+		return turtle.URI{}, err
+	}
+	uri := turtle.ParseURI(string(val))
+	db.uriCache[hash] = uri
+	return uri, nil
+}
+
+func (db *DB) MustGetURI(hash Key) turtle.URI {
+	if hash == emptyKey {
+		return turtle.URI{}
+	}
+	uri, err := db.GetURI(hash)
+	if err != nil {
+		log.Error(errors.Wrapf(err, "Could not get URI for %v", hash))
+		return turtle.URI{}
+	}
+	return uri
+}
+
+func (db *DB) MustGetURIStringHash(hash string) turtle.URI {
+	var c Key
+	copy(c[:], []byte(hash))
+	return db.MustGetURI(c)
+}
+
+func (db *DB) GetEntity(uri turtle.URI) (*Entity, error) {
+	hash, err := db.GetHash(uri)
+	if err != nil {
+		return nil, err
+	}
+	return db.GetEntityFromHash(hash)
+}
+
+func (db *DB) GetEntityFromHash(hash Key) (*Entity, error) {
+	db.eocLock.RLock()
+	if ent, found := db.entityObjectCache[hash]; found {
+		db.eocLock.RUnlock()
+		return ent, nil
+	}
+	db.eocLock.RUnlock()
+	db.eocLock.Lock()
+	defer db.eocLock.Unlock()
+	bytes, err := db.graphDB.Get(hash[:], nil)
+	if err != nil {
+		return nil, errors.Wrapf(err, "Could not get Entity from graph for %s", db.MustGetURI(hash))
+	}
+	ent := NewEntity()
+	_, err = ent.UnmarshalMsg(bytes)
+	db.entityObjectCache[hash] = ent
+	return ent, err
+}
+
+func (db *DB) MustGetEntityFromHash(hash Key) *Entity {
+	e, err := db.GetEntityFromHash(hash)
+	if err != nil {
+		log.Error(errors.Wrap(err, "Could not get entity"))
+		return nil
+	}
+	return e
+}
+
+func (db *DB) MustGetEntityIndexFromHash(hash Key) *EntityExtendedIndex {
+	e, err := db.GetEntityIndexFromHash(hash)
+	if err != nil {
+		log.Error(errors.Wrap(err, "Could not get entity index"))
+		return nil
+	}
+	return e
+}
+
+func (db *DB) DumpEntity(ent *Entity) {
+	fmt.Println("DUMPING", db.MustGetURI(ent.PK))
+	for edge, list := range ent.OutEdges {
+		fmt.Printf(" OUT: %s \n", db.MustGetURIStringHash(edge).Value)
+		for _, l := range list {
+			fmt.Printf("     -> %s\n", db.MustGetURI(l).Value)
+		}
+	}
+	for edge, list := range ent.InEdges {
+		fmt.Printf(" In: %s \n", db.MustGetURIStringHash(edge).Value)
+		for _, l := range list {
+			fmt.Printf("     <- %s\n", db.MustGetURI(l).Value)
+		}
+	}
+}
+
+func (db *DB) GetEntityTx(graphtx *leveldb.Transaction, uri turtle.URI) (*Entity, error) {
+	var entity = NewEntity()
+	hash, err := db.GetHash(uri)
+	if err != nil {
+		return nil, err
+	}
+	bytes, err := graphtx.Get(hash[:], nil)
+	if err != nil {
+		return nil, err
+	}
+	_, err = entity.UnmarshalMsg(bytes)
+	if err != nil {
+		return nil, err
+	}
+	return entity, nil
+}
+
+func (db *DB) GetEntityFromHashTx(graphtx *leveldb.Transaction, hash Key) (*Entity, error) {
+	bytes, err := graphtx.Get(hash[:], nil)
+	if err != nil {
+		return nil, err
+	}
+	ent := NewEntity()
+	_, err = ent.UnmarshalMsg(bytes)
+	return ent, err
+}
+
+func (db *DB) GetEntityIndexFromHashTx(extendtx *leveldb.Transaction, hash Key) (*EntityExtendedIndex, error) {
+	bytes, err := extendtx.Get(hash[:], nil)
+	if err != nil {
+		return nil, err
+	}
+	ent := NewEntityExtendedIndex()
+	_, err = ent.UnmarshalMsg(bytes)
+	return ent, err
+}
+
+func (db *DB) GetEntityIndexFromHash(hash Key) (*EntityExtendedIndex, error) {
+	db.eicLock.RLock()
+	if ent, found := db.entityIndexCache[hash]; found {
+		db.eicLock.RUnlock()
+		return ent, nil
+	}
+	db.eicLock.RUnlock()
+	db.eicLock.Lock()
+	defer db.eicLock.Unlock()
+	bytes, err := db.extendedDB.Get(hash[:], nil)
+	if err != nil && err != leveldb.ErrNotFound {
+		return nil, errors.Wrapf(err, "Could not get EntityIndex from graph for %s", db.MustGetURI(hash))
+	} else if err == leveldb.ErrNotFound {
+		db.entityIndexCache[hash] = nil
+		return nil, nil
+	}
+	ent := NewEntityExtendedIndex()
+	_, err = ent.UnmarshalMsg(bytes)
+	db.entityIndexCache[hash] = ent
+	return ent, err
 }
